net: tcp: Add tcp options length
The tcphdr->offset was not set when tcp options were added.


Former-commit-id: 07b8e4cb01dcb364752b83e63e371b648e200aba

+++ subsys/net/ip/tcp.c
@@ -17,7 +17,7 @@ DECL|ip_max_packet_len|function|static inline size_t ip_max_packet_len(struct in
 DECL|ip_max_packet_len|macro|ip_max_packet_len
 DECL|is_6lo_technology|macro|is_6lo_technology
 DECL|net_tcp_ack_received|function|void net_tcp_ack_received(struct net_context *ctx, uint32_t ack)
-DECL|net_tcp_add_options|function|static inline int net_tcp_add_options(struct net_buf *header, size_t len, void *data)
+DECL|net_tcp_add_options|function|static inline uint8_t net_tcp_add_options(struct net_buf *header, size_t len, void *data)
 DECL|net_tcp_alloc|function|struct net_tcp *net_tcp_alloc(struct net_context *context)
 DECL|net_tcp_change_state|function|void net_tcp_change_state(struct net_tcp *tcp, enum net_tcp_state new_state)
 DECL|net_tcp_foreach|function|void net_tcp_foreach(net_tcp_cb_t cb, void *user_data)
